+++ airsen-backend/src/main/java/fr/airsen/api/dto/response/WeatherExportDTO.java
@@ -10,9 +10,9 @@ import java.time.LocalDateTime;
 public record WeatherExportDTO(
     LocalDateTime observationDate,
     Double temperature,
-    Double humidity,
+    Integer humidity,
     Double windSpeed,
-    Double windDirection,
+    Integer windDirection,
     Integer weatherCode,
     LocalDateTime createdAt
 ) {}
